+++ MOP/src/GameObjects/WorldObjects.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Generic;
 using UnityEngine;
 
 namespace MOP
@@ -17,31 +14,31 @@ namespace MOP
 
         public void Add(string gameObjectName, int distance = 200, bool rendererOnly = false)
         {
-            if (GameObject.Find(gameObjectName))
-            {
-                worldObject.Add(new WorldObject(gameObjectName, distance, rendererOnly));
-            }
-            else
+            if (GameObject.Find(gameObjectName) == null)
             {
                 MSCLoader.ModConsole.Print("[MOP] Couldn't find " + gameObjectName + ".");
+                return;
             }
+
+            worldObject.Add(new WorldObject(gameObjectName, distance, rendererOnly));
         }
 
         public void Add(string gameObjectName, bool awayFromHouse, bool rendererOnly = false)
         {
-            if (GameObject.Find(gameObjectName))
-            {
-                worldObject.Add(new WorldObject(gameObjectName, awayFromHouse, rendererOnly));
-            }
-            else
+            if (GameObject.Find(gameObjectName) == null)
             {
                 MSCLoader.ModConsole.Print("[MOP] Couldn't find " + gameObjectName + ".");
+                return;
             }
+
+            worldObject.Add(new WorldObject(gameObjectName, awayFromHouse, rendererOnly));
         }
 
         public WorldObject Get(int i)
         {
             return worldObject[i];
         }
+
+        public int Count => worldObject.Count;
     }
 }