+++ MOP/src/Occlusion/OcclusionObject.cs
@@ -90,7 +90,7 @@ namespace MOP
 
         public void Toggle()
         {
-            PlayerDistance = Vector3.Distance(WorldManager.instance.player.position, transform.position);
+            PlayerDistance = Vector3.Distance(WorldManager.instance.Player.position, transform.position);
             bool visible = (IsVisible) || (PlayerDistance < MopSettings.MinOcclusionDistance);
 
             for (int i = 0; i < Renderers.Count; i++)